сделал все запросы и ответы через dto

+++ src/main/java/com/example/cafebills/service/BillService.java
@@ -1,40 +1,47 @@
 package com.example.cafebills.service;
 
-import com.example.cafebills.dto.ProductDto;
-import com.example.cafebills.entity.Product;
+import com.example.cafebills.dto.BillRequestDto;
+import com.example.cafebills.dto.BillResponseDto;
+import com.example.cafebills.entity.Bill;
+import com.example.cafebills.mapper.BillDtoMapper;
+import com.example.cafebills.repository.BillRepository;
+import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import com.example.cafebills.dto.BillDto;
-import com.example.cafebills.entity.Bill;
-import com.example.cafebills.repository.BillRepository;
 
 import java.util.List;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class BillService {
 
     private final BillRepository billRepository;
+    private final BillDtoMapper billDtoMapper;
 
-    public Bill createBill(double price, List<Product> productList) {
-        Bill bill = new Bill();
-        bill.setPrice(price);
-        bill.setProductList(productList);
-        return billRepository.save(bill);
+    public BillResponseDto createBill(BillRequestDto requestDto) {
+        Bill bill = billDtoMapper.toEntity(requestDto);
+        Bill savedBill = billRepository.save(bill);
+        return billDtoMapper.toDto(savedBill);
     }
 
-    public Bill getBillById(UUID billId) {
-        return billRepository.findById(billId).orElse(null);
+    public BillResponseDto getBillById(@NotNull UUID billId) {
+        Bill bill = billRepository.findById(billId).orElse(null);
+        return (bill != null) ? billDtoMapper.toDto(bill) : null;
     }
 
-    public List<Bill> getAllBills() {
-        return billRepository.findAll();
+    public List<BillResponseDto> getAllBills() {
+        List<Bill> bills = billRepository.findAll();
+        return bills.stream()
+                .map(billDtoMapper::toDto)
+                .toList();
     }
 
-    public void deleteBill(UUID billId) {
+    public boolean deleteBill(@NotNull UUID billId) {
+        if (!billRepository.existsById(billId)) {
+            return false;
+        }
         billRepository.deleteById(billId);
+        return true;
     }
 }
-
