+++ src/main/java/com/example/cafebills/service/ProductService.java
@@ -1,10 +1,13 @@
 package com.example.cafebills.service;
 
+import com.example.cafebills.dto.ProductRequestDto;
+import com.example.cafebills.dto.ProductResponseDto;
 import com.example.cafebills.entity.Product;
+import com.example.cafebills.mapper.ProductDtoMapper;
 import com.example.cafebills.repository.ProductRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
 import java.util.UUID;
 
@@ -13,23 +16,31 @@ import java.util.UUID;
 public class ProductService {
 
     private final ProductRepository productRepository;
+    private final ProductDtoMapper productDtoMapper;
 
-
-    public Product createProduct(String name) {
-        Product product = new Product();
-        product.setName(name);
-        return productRepository.save(product);
+    public ProductResponseDto createProduct(ProductRequestDto requestDto) {
+        Product product = productDtoMapper.toEntity(requestDto);
+        Product savedProduct = productRepository.save(product);
+        return productDtoMapper.toDto(savedProduct);
     }
 
-    public Product getProductById(UUID productId) {
-        return productRepository.findById(productId).orElse(null);
+    public ProductResponseDto getProductById(UUID productId) {
+        Product product = productRepository.findById(productId).orElse(null);
+        return (product != null) ? productDtoMapper.toDto(product) : null;
     }
 
-    public List<Product> getAllProducts() {
-        return productRepository.findAll();
+    public List<ProductResponseDto> getAllProducts() {
+        List<Product> products = productRepository.findAll();
+        return products.stream()
+                .map(productDtoMapper::toDto)
+                .toList();
     }
 
-    public void deleteProduct(UUID productId) {
+    public boolean deleteProduct(UUID productId) {
+        if (!productRepository.existsById(productId)) {
+            return false;
+        }
         productRepository.deleteById(productId);
+        return true;
     }
 }
